Move http-api clients to rl-client
