+++ runelite-client/src/main/java/net/runelite/client/ClientSessionManager.java
@@ -38,7 +38,6 @@ import net.runelite.api.GameState;
 import net.runelite.client.eventbus.Subscribe;
 import net.runelite.client.events.ClientShutdown;
 import net.runelite.client.util.RunnableExceptionLogger;
-import okhttp3.OkHttpClient;
 
 @Singleton
 @Slf4j
@@ -54,11 +53,11 @@ public class ClientSessionManager
 	@Inject
 	ClientSessionManager(ScheduledExecutorService executorService,
 		@Nullable Client client,
-		OkHttpClient okHttpClient)
+		SessionClient sessionClient)
 	{
 		this.executorService = executorService;
 		this.client = client;
-		this.sessionClient = new SessionClient(okHttpClient);
+		this.sessionClient = sessionClient;
 	}
 
 	public void start()
