+++ runelite-client/src/main/java/net/runelite/client/game/NPCManager.java
@@ -32,21 +32,18 @@ import javax.annotation.Nullable;
 import javax.inject.Inject;
 import javax.inject.Singleton;
 import lombok.extern.slf4j.Slf4j;
-import net.runelite.http.api.npc.NpcInfo;
-import net.runelite.http.api.npc.NpcInfoClient;
-import okhttp3.OkHttpClient;
 
 @Singleton
 @Slf4j
 public class NPCManager
 {
-	private final OkHttpClient okHttpClient;
+	private final NpcInfoClient npcInfoClient;
 	private Map<Integer, NpcInfo> npcMap = Collections.emptyMap();
 
 	@Inject
-	private NPCManager(OkHttpClient okHttpClient, ScheduledExecutorService scheduledExecutorService)
+	private NPCManager(NpcInfoClient npcInfoClient, ScheduledExecutorService scheduledExecutorService)
 	{
-		this.okHttpClient = okHttpClient;
+		this.npcInfoClient = npcInfoClient;
 		scheduledExecutorService.execute(this::loadNpcs);
 	}
 
@@ -67,7 +64,7 @@ public class NPCManager
 	{
 		try
 		{
-			npcMap = new NpcInfoClient(okHttpClient).getNpcs();
+			npcMap = npcInfoClient.getNpcs();
 		}
 		catch (IOException e)
 		{
