+++ runelite-client/src/main/java/net/runelite/client/game/NpcInfo.java
@@ -22,7 +22,7 @@
  * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
  * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-package net.runelite.http.api.npc;
+package net.runelite.client.game;
 
 import lombok.Data;
