+++ runelite-client/src/main/java/net/runelite/client/game/WorldClient.java
@@ -1,5 +1,6 @@
 /*
- * Copyright (c) 2019, Adam <Adam@sigterm.info>
+ * Copyright (c) 2017, Adam <Adam@sigterm.info>
+ * Copyright (c) 2018, Lotto <https://github.com/devLotto>
  * All rights reserved.
  *
  * Redistribution and use in source and binary forms, with or without
@@ -22,72 +23,55 @@
  * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
  * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-package net.runelite.http.api.ge;
+package net.runelite.client.game;
 
-import com.google.gson.Gson;
+import com.google.gson.JsonParseException;
 import java.io.IOException;
-import java.util.UUID;
+import java.io.InputStream;
+import java.io.InputStreamReader;
+import java.nio.charset.StandardCharsets;
 import lombok.RequiredArgsConstructor;
-import lombok.Setter;
 import lombok.extern.slf4j.Slf4j;
 import net.runelite.http.api.RuneLiteAPI;
-import static net.runelite.http.api.RuneLiteAPI.JSON;
-import okhttp3.Call;
-import okhttp3.Callback;
+import net.runelite.http.api.worlds.WorldResult;
 import okhttp3.HttpUrl;
 import okhttp3.OkHttpClient;
 import okhttp3.Request;
-import okhttp3.RequestBody;
 import okhttp3.Response;
 
 @Slf4j
 @RequiredArgsConstructor
-public class GrandExchangeClient
+public class WorldClient
 {
-	private static final Gson GSON = RuneLiteAPI.GSON;
-
 	private final OkHttpClient client;
+	private final HttpUrl apiBase;
 
-	@Setter
-	private UUID uuid;
-	@Setter
-	private String machineId;
-
-	public void submit(GrandExchangeTrade grandExchangeTrade)
+	public WorldResult lookupWorlds() throws IOException
 	{
-		final HttpUrl url = RuneLiteAPI.getApiBase().newBuilder()
-			.addPathSegment("ge")
+		HttpUrl url = apiBase.newBuilder()
+			.addPathSegment("worlds.js")
 			.build();
 
-		Request.Builder builder = new Request.Builder();
-		if (uuid != null)
-		{
-			builder.header(RuneLiteAPI.RUNELITE_AUTH, uuid.toString());
-		}
-		if (machineId != null)
-		{
-			builder.header(RuneLiteAPI.RUNELITE_MACHINEID, machineId);
-		}
+		log.debug("Built URI: {}", url);
 
-		Request request = builder
-			.post(RequestBody.create(JSON, GSON.toJson(grandExchangeTrade)))
+		Request request = new Request.Builder()
 			.url(url)
 			.build();
 
-		client.newCall(request).enqueue(new Callback()
+		try (Response response = client.newCall(request).execute())
 		{
-			@Override
-			public void onFailure(Call call, IOException e)
+			if (!response.isSuccessful())
 			{
-				log.debug("unable to submit trade", e);
+				log.debug("Error looking up worlds: {}", response);
+				throw new IOException("unsuccessful response looking up worlds");
 			}
 
-			@Override
-			public void onResponse(Call call, Response response)
-			{
-				log.debug("Submitted trade");
-				response.close();
-			}
-		});
+			InputStream in = response.body().byteStream();
+			return RuneLiteAPI.GSON.fromJson(new InputStreamReader(in, StandardCharsets.UTF_8), WorldResult.class);
+		}
+		catch (JsonParseException ex)
+		{
+			throw new IOException(ex);
+		}
 	}
 }
