+++ runelite-client/src/main/java/net/runelite/client/game/WorldService.java
@@ -33,6 +33,7 @@ import java.util.concurrent.TimeUnit;
 import java.util.concurrent.TimeoutException;
 import javax.annotation.Nullable;
 import javax.inject.Inject;
+import javax.inject.Named;
 import javax.inject.Singleton;
 import lombok.extern.slf4j.Slf4j;
 import net.runelite.api.Client;
@@ -41,8 +42,8 @@ import net.runelite.client.eventbus.EventBus;
 import net.runelite.client.events.WorldsFetch;
 import net.runelite.client.util.RunnableExceptionLogger;
 import net.runelite.http.api.worlds.World;
-import net.runelite.http.api.worlds.WorldClient;
 import net.runelite.http.api.worlds.WorldResult;
+import okhttp3.HttpUrl;
 import okhttp3.OkHttpClient;
 
 @Singleton
@@ -61,11 +62,11 @@ public class WorldService
 
 	@Inject
 	private WorldService(Client client, ScheduledExecutorService scheduledExecutorService, OkHttpClient okHttpClient,
-		EventBus eventBus)
+		@Named("runelite.api.base") HttpUrl apiBase, EventBus eventBus)
 	{
 		this.client = client;
 		this.scheduledExecutorService = scheduledExecutorService;
-		this.worldClient = new WorldClient(okHttpClient);
+		this.worldClient = new WorldClient(okHttpClient, apiBase);
 		this.eventBus = eventBus;
 
 		scheduledExecutorService.scheduleWithFixedDelay(RunnableExceptionLogger.wrap(this::tick), 0, WORLD_FETCH_TIMER, TimeUnit.MINUTES);
