+++ runelite-client/src/main/resources/net/runelite/client/runelite.properties
@@ -13,3 +13,7 @@ runelite.jav_config_backup=https://static.runelite.net/jav_config.ws
 runelite.pluginhub.url=https://repo.runelite.net/plugins
 runelite.pluginhub.version=${project.version}
 runelite.imgur.client.id=30d71e5f6860809
+runelite.api.base=https://api.runelite.net/runelite-${project.version}
+runelite.session=https://api.runelite.net/session
+runelite.static.base=https://static.runelite.net
+runelite.ws=https://api.runelite.net/ws
