+++ runelite-client/src/test/java/net/runelite/client/plugins/xptracker/XpTrackerPluginTest.java
@@ -39,7 +39,6 @@ import net.runelite.client.game.NPCManager;
 import net.runelite.client.game.SkillIconManager;
 import net.runelite.client.ui.ClientToolbar;
 import net.runelite.client.ui.overlay.OverlayManager;
-import net.runelite.http.api.xp.XpClient;
 import static org.junit.Assert.assertEquals;
 import org.junit.Before;
 import org.junit.Test;
